Define Fixes, More Warning Fixes
It's slowly coming together...

+++ CvGameCoreDLL_Expansion2/CvPlayer.cpp
@@ -17326,7 +17326,7 @@ int CvPlayer::GetScienceFromOtherPlayersTimes100() const
 		{
 			if (GET_PLAYER(ePlayer).GetMinorCivAI()->IsAllies(GetID()))
 			{
-				switch (GetCurrentEra())
+				switch ((int)GetCurrentEra())
 				{
 					case 0:
 						iScienceFromPlayer = 0;
